refactor CurrentCustomDatasetsBuilder and IndexDatasetsBuilder
