+++ packages/suite-base/src/panels/Plot/builders/IndexDatasetsBuilder.ts
@@ -5,11 +5,8 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { ChartDataset } from "chart.js";
-
 import { filterMap } from "@lichtblick/den/collection";
-import { MessagePath } from "@lichtblick/message-path";
-import { Immutable, Time, MessageEvent } from "@lichtblick/suite";
+import { Immutable } from "@lichtblick/suite";
 import { simpleGetMessagePathDataItems } from "@lichtblick/suite-base/components/MessagePathSyntax/simpleGetMessagePathDataItems";
 import { Bounds1D } from "@lichtblick/suite-base/components/TimeBasedChart/types";
 import { PlayerState } from "@lichtblick/suite-base/players/types";
@@ -22,26 +19,17 @@ import {
   SeriesConfigKey,
   SeriesItem,
 } from "./IDatasetsBuilder";
+import {
+  CurrentFrameSeriesItem,
+  buildViewportDatasets,
+  lastMatchingTopic,
+  setSeries,
+} from "./utils";
 import { MATH_FUNCTIONS } from "../constants";
-import { Dataset } from "../types";
-import { getChartValue, isChartValue, Datum } from "../utils/datum";
-
-type DatumWithReceiveTime = Datum & {
-  receiveTime: Time;
-};
-
-type IndexDatasetsSeries = {
-  configIndex: number;
-  enabled: boolean;
-  messagePath: string;
-  parsed: Immutable<MessagePath>;
-  dataset: ChartDataset<"scatter", DatumWithReceiveTime[]>;
-};
-
-const emptyPaths = new Set<string>();
+import { getChartValue, isChartValue } from "../utils/datum";
 
 export class IndexDatasetsBuilder implements IDatasetsBuilder {
-  #seriesByKey = new Map<SeriesConfigKey, IndexDatasetsSeries>();
+  #seriesByKey = new Map<SeriesConfigKey, CurrentFrameSeriesItem>();
 
   #range?: Bounds1D;
 
@@ -104,38 +92,7 @@ export class IndexDatasetsBuilder implements IDatasetsBuilder {
   }
 
   public setSeries(series: Immutable<SeriesItem[]>): void {
-    // Make a new map so we drop series which are no longer present
-    const newSeries = new Map();
-
-    for (const item of series) {
-      let existingSeries = this.#seriesByKey.get(item.key);
-      existingSeries ??= {
-        configIndex: item.configIndex,
-        enabled: item.enabled,
-        messagePath: item.messagePath,
-        parsed: item.parsed,
-        dataset: {
-          data: [],
-        },
-      };
-
-      existingSeries.configIndex = item.configIndex;
-      existingSeries.enabled = item.enabled;
-      existingSeries.dataset = {
-        ...existingSeries.dataset,
-        borderColor: item.color,
-        showLine: item.showLine,
-        fill: false,
-        borderWidth: item.lineSize,
-        pointRadius: item.lineSize * 1.2,
-        pointHoverRadius: 3,
-        pointBackgroundColor: item.showLine ? item.contrastColor : item.color,
-        pointBorderColor: "transparent",
-      };
-
-      newSeries.set(item.key, existingSeries);
-    }
-    this.#seriesByKey = newSeries;
+    this.#seriesByKey = setSeries(this.#seriesByKey, series);
   }
 
   // We don't use the viewport because we do not do any downsampling on the assumption that
@@ -143,14 +100,7 @@ export class IndexDatasetsBuilder implements IDatasetsBuilder {
   //
   // If that assumption changes then downsampling can be revisited.
   public async getViewportDatasets(): Promise<GetViewportDatasetsResult> {
-    const datasets: Dataset[] = [];
-    for (const series of this.#seriesByKey.values()) {
-      if (series.enabled) {
-        datasets[series.configIndex] = series.dataset;
-      }
-    }
-
-    return { datasetsByConfigIndex: datasets, pathsWithMismatchedDataLengths: emptyPaths };
+    return buildViewportDatasets(this.#seriesByKey);
   }
 
   public async getCsvData(): Promise<CsvDataset[]> {
@@ -168,15 +118,4 @@ export class IndexDatasetsBuilder implements IDatasetsBuilder {
 
     return datasets;
   }
-}
-
-function lastMatchingTopic(msgEvents: Immutable<MessageEvent[]>, topic: string) {
-  for (let i = msgEvents.length - 1; i >= 0; --i) {
-    const msgEvent = msgEvents[i]!;
-    if (msgEvent.topic === topic) {
-      return msgEvent;
-    }
-  }
-
-  return undefined;
 }
